Fix back button problems and map loading
I was able to fix the back button problems, by refering to the correct
screen element in the on back cmd. I was also able to solve the map
loading problem, because it seems that I only loaded the map when a user
navigated from the list page to the detail page, and forgot to reset the
map when the detail page was navigated to by the productsdetail page.