Fix back nav problems and education loading
We were having problems getting back to the correct detail page from the info
page because we were caling the detail page. Also we needed the story
store to be populated with the correct data, and we needed to display that
data correctly. We will now focus on updating the app from SeaGrant_Proto
to WhatsFresh. Then we can work on structuring the education data better.