Redesign info page
I have been working on making the info page contain a picture, historical
data and a list of other items that pertain to the product that the user
wants to learn more about. I was able to get the list and picture to
resize dynamicly, but not the historical information.