Remove the numberOfVendors() function and supporting data structures.
Prior to this commit the numberOfVendors function was making a second
copy of store data. This copy was used to define the map markers on
the Map.js page. This commit connects the logic defining map markers
directly to the vendor store such that we no longer need to maintain
the extra structures used to implement numberOfVendors().

Conflicts:
	app/controller/List.js