Redesign the Educational information
I renamed the list items to display all of the necisary educational data,
and made sure the image was at the top of the page with historical
information below. We still need to do a bit more css so that the
historical data is displayed in such a way that you can read it all. We
probably need to make it so that a user can scroll through the historical
data. I also dealt with the case when a product didn't have any
educational data.