Introduce Search.js util code from working-waterfronts-mobile
This commit begins the process of porting the Search util from Working
Waterfronts to the Whats Fresh project. In this commit, the Jasmine
test data collection, Search.js and its corresponding test code is
copied into the repo. SpecRunner.html is updated to include these
references.

This commit does not begin the work of adjusting this util to also
take product filtering into account.

This commit also accidentally includes an off-topic edit: Prior to
this commit, the distanceSpec test had a syntax error preventing
successful testing. This commit fixes a single curly brace and
resolves the test.