Fix infowindow navigation issues
I realized that the infowindow button was not being populated with the
data it needed to navigate to the detail page, so we made that possible by
setting WhatsFresh.storeItem outside of the added listener for the map
markers. This way the list would be able to set the data the infowindow
button needed, when it opened the infowindow corresponding to the list
item. Also had to fix an issue where the previously highlighted list item
was not being changed, so we set it to null when the user navigates back
from the detail page.