Port Search.js util to What's Fresh (complete)
This commit finishes updating the Search.js util and its related test
code/test data to match its new scope in whats-fresh-mobile. Search.js
can now filter by products.

This commit includes changes to test data to properly test distance
filtering and product filtering.

Currently, this util correctly filters vendors in the vendor store but
is not connected to the rest of the app. The next steps for this
feature will include:

- Replace filter helper functions in List.js with this util
- Implement cross-filtering on Products depending on the state of this search