Add Crossfilter util to handle propegate Search util results
This commit introduces app/util/Crossfilter.js and corresponding tests
to the project. This util is designed to filter the product store
based on the results of the filtering process that is currently
applied to the vendor store.

The project currently builds, test pass and the app functions as
before on dev branch. However, neither the Search util nor the
Crossfilter util have been integrated into the app's execution path.