Remove extra buttons and format
I got rid of the buttons Detail, Info, and Specific as well as their
listeners, function calls, and functions. I also fixed the map, because it
was showing the map on 100% of the page, but only 50% of the map was
showing because of the list. Thus we made the map 50% of the height.