Wrap Homescreen elements in Ext fieldsets
Prior to this commit, the home screen did not use fieldsets to
properly format and separate discrete portions of the home screen's
search features. This commit wraps these groups of screen elements
(such as the location and product dropdowns) in their own fieldsets to
break them into their own scope visually and in the view's code.