Align searchSummaryTpl text to center on Home screen
This commit introduces a clause in the app.scss file to align the
searchSummaryTpl text to the center of the "vendnum" div.

This commit does not alter the min-height of the vendnum div to
prevent interface "wobble" when the status bar needs to wrap its text.
Future iterations of this app will need to clean and organize the
naming conventions of the UI code as well as the corresponding CSS
properties.