Relocate the images/ folder to resources/images/
Prior to this commit, the images folder was located at the root of the
project. As the images folder only contains information relevant to
rendering google maps nodes inside of the app's feature code, this
commit relocates it to the resources folder and updates the references
in app/controller/List.js to ensure that the map can still render.

Future work in this branch should include:
- Removing docs resources out of the root of the project
- Removing npm resources out of the root of the project
- Reinitializing phonegap resources