Update phonegap config.xml and relocate it to phonegap root
This commit updates the phonegap config.xml file to it's new root
location in the phonegap folder. Prior to this commit, the config.xml
file was located in the phonegap/www/ folder. The copy of config.xml
at this location is a file generated by Sencha before Sencha calls
phonegap's build process. By moving the current copy of config.xml to
its new location, we can safely separate configuration and build
results.