Repair .sencha config and version information
This commit updates and repairs the .sencha configuration information
to the latest version of Sencha Cmd. Prior to this commit, running
Sencha's app build command would create nested resource/ directories,
making it impossible to consistently path to app resources in phonegap
apps. This commit introduces the results of the `sencha app upgrade
-noframework` command which repairs Sencha's nested resources/ build
behavior

This commit also updates the .gitignore file to exclude the
.sencha_backup directory that is created by `sencha app upgrade`.

This commit also updates build.xml and index.html to reflect the
changes to .sencha by `sencha app upgrade`