Update phonegap/config.xml with labels, iOS geolocation feature
Prior to this commit, phonegap's config.xml file was outdated: it did
not reflect the new user-facing title of the app: Oregon's Catch, it
did not reflect the app id, and it did not contain plugin information
for using phonegap's geolocation plugin in iOS devices.

This commit updates these metadata in phonegap/config.xml where it
will be copied into the correct locations by Phonegap CLI and Sencha
Cmd build functions.