Repair device callback errors when deployed to device
Prior to this commit, devicePositionCallback(position) and
devicePositionFailure were unable to properly use scope when the app
was deployed to android test devices. This commit repairs this
functionality by in-lining these callback functions inside of
watchPosition() to allow them to correctly refer to objects inside the
controller scope.

These changes also implied repairs to the onSetDistance() function as
its functionality was buried inside the callback functions that were
rewritten.

As of this commit:

1) Device geolocation is still broken. The phonegap/cordova plugin does
not callback OR fail in the event location services are just turned
off.

2) The Product map/list screen does not properly display the list of
searched inventory items. Something in the new device location code
breaks this screen when the device's geolocation callback fires.