Repair search summary behavior when device toggles GPS
Prior to this commit, toggling device location off did not update the
search summary. This commit adds code to filter and update accordingly
when the device's location service is turned off.

This fix also addresses a first-time-search bug where the user is
unable to filter by device location: the Search.options.distance field
is now set every time device location is toggled on.

This commit also removes the WhatsFresh.userLoc global variable by
checking for the toggle's state when the search summary is built.