Introduce ad-hoc testing for null device position
Prior to this commit, the app was unable to fail gracefully when
deployed to a device with location services turned off. This is
because the phonegap/cordova geolocation plugin is not null safe.

This commit introduces an Ext.Function.defer() test to the
onSetUseLocation function to check for nulls in the incoming stream of
device position data. If the check encounters a null, the app will now
assume that geolocation permissions are correctly set but that
location services are turned off and should be enabled.