Fix callback reference to populatePstore() and refactor populatePstore() for clarity.
Prior to this commit, sorting search results by product would fail as
soon as the Geolocation service callback was fired. This commit
repairs the call to populatePstore() in the Geolocation callback to
allow sorting to function properly.

This commit refactors populatePstore() for clarity, but does not fix
code cleanliness. populatePstore() currently has a TODO indicating how
to remove the unused WhatsFresh.use global variable.