Update phonegap/config.xml and hooks for ios builds
Prior to this commit, whats-fresh-mobile was unable to compile ios
projects. This commit fixes escape characters in the config.xml name
field, links splash screens and updates hooks for phonegap ios
builds. As of this commit, `sencha app build ios` generates an xcode
project that will deploy to a simulator.

This commit also discovers that the core app functionality does not
work on iOS. This bugfix should be addressed in a new branch.