Relocate Google maps resource link from index.html to app.json
Prior to this commit, the app was unable to launch on iOS devices due
to an error loading/connecting to the Google Maps javascript api
resources. This commit moves the reference to the Google Maps api from
the index.html file to the app.json `resources` object. As of this
commit, the app is able to launch on iOS devices *if* it is able to
connect to the internet.