Use phonegap geolocation and fix error msg
We switched the geolocation back to using phonegap geolocation, so that
geolocation would work ok with all of the different android versions. Then
we added a line to the error message that comes up when the user trys to
search by location, but can't. The added msg lets them know that they need
to restart the app after they have turned location services on for their
device.