Remove empty source/navpage from toctree and begin navigation instructions.
Prior to this commit, the toctree included an empty source/navpage
file and did not include step-by-step high-level instructions for
using/navigating the app. This commit removes the source/navpage
reference and begins the task of writing step-by-step high level
instructions on how to use the app.

Next commits should include high level discussion of how to use each
of the app's pages.