Create new "errorStatus" display to show location errors
Prior to this commit, location error messages were clobbering one
another other on the home screen. This commit introduces code to the
home view and the list controller to provide a write-safe location for
error messages to display. The home view now contains an "errorStatus"
object that can be hidden or shown depending on the error message. The
list controller now contains code to write location errors to the
errorStatus object and hide it when appropriate.